Removed extensions
